app/requirements.txt
psycopg2-binary==2.9.9
faker==24.8.0